reorganize to try and split things up just a little bit
