+++ lib/saluki-core/src/runtime/supervisor.rs
@@ -308,10 +308,9 @@ where
 /// # Instrumentation
 ///
 /// Supervisors automatically create their own allocation group
-/// ([`TrackingAllocator`][resource_accounting::allocator::TrackingAllocator]), which is used to track both the memory
-/// usage of the supervisor itself and its children. Additionally, individual worker processes are wrapped in a
-/// dedicated [`tracing::Span`] to allow tracing the causal relationship between arbitrary code and the worker executing
-/// it.
+/// ([`TrackingAllocator`][resource_accounting::TrackingAllocator]), which is used to track both the memory usage of the
+/// supervisor itself and its children. Additionally, individual worker processes are wrapped in a dedicated
+/// [`tracing::Span`] to allow tracing the causal relationship between arbitrary code and the worker executing it.
 ///
 /// # Restart Strategies
 ///
